Improve auth
