+++ server/api.js
@@ -1,4 +1,4 @@
-const verifyToken = require("./auth/auth");
+const auth = require("./auth/auth.middleware");
 const express = require("express");
 const router = express.Router();
 
@@ -7,7 +7,7 @@ const titulosController = require('./titulos/titulos.controller');
 const cowboysRoute = require('./cowboys/cowboys.route');
 const putosRoute = require('./putos/putos.route');
 
-router.use(verifyToken);
+router.use(auth);
 
 router.use('/cowboys', cowboysRoute);
 router.use('/putos', putosRoute); // TODO: remove
@@ -16,6 +16,7 @@ router.use('/putos', putosRoute); // TODO: remove
 router.get( '/cowboys/:idCowboy/titulos', titulosController.getCowboyTitulos);
 // Solicitud `POST /cowboys/idCowboy/titulos` para crear un titulo al un cowboy por id
 router.post('/cowboys/:idCowboy/titulos', titulosController.createCowboyTitulo);
+
 // Solicitud `GET /titulos para obtener` todos los títulos
 router.get('/titulos', titulosController.getAllTitulos);
 // Solicitud `DELETE /titulos/:idTitulo` para obtener todos los títulos
