+++ server/index.js
@@ -9,32 +9,38 @@ const RateLimit = require("express-rate-limit");
 const apiRoute = require('./api');
 
 const app = express();
-const PORT = 3001;
+const PORT = 3000;
 
 const web = express.static('frontend/dist');
 
 // Set up rate limiter: maximum of twenty requests per minute
 const limiter = RateLimit({
-  windowMs: 1 * 60 * 1000, // 1 minute
+  windowMs: 15 * 60 * 1000, // 15 minute
   max: 100,
 });
 
+
+const hlmetHeaders = helmet.contentSecurityPolicy({
+  directives: {
+    "script-src": ["'self'", "fonts.googleapis.com"],
+  },
+});
+
+// gzip compression
 app.use(compression());
-app.use(
-  helmet.contentSecurityPolicy({
-    directives: {
-      "script-src": ["'self'", "fonts.googleapis.com"],
-    },
-  }),
-);
+// set security HTTP headers
+app.use(hlmetHeaders);
+//  apply to all requests
 app.use(limiter);
+// enable cors
 app.use(cors());
+// parse json request body
 app.use(express.json());
+// static files
 app.use(express.static('frontend/dist'));
+
 app.use('/', web);
-
 app.use('/api', apiRoute);
-
 app.use('*', web);
 
 // Start the server
